@@ -11,10 +11,10 @@
     "test": "jest"
   },
   "devDependencies": {
-    "autoprefixer": "^8.2.0",
+    "autoprefixer": "^8.3.0",
     "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.0",
-    "babel-eslint": "^8.1.2",
+    "babel-core": "^6.26.2",
+    "babel-eslint": "^8.2.3",
     "babel-jest": "^22.4.3",
     "babel-loader": "^7.1.2",
     "babel-plugin-transform-class-properties": "^6.24.1",
@@ -26,7 +26,7 @@
     "eslint": "^4.19.1",
     "eslint-config-airbnb": "^16.0.0",
     "eslint-config-prettier": "^2.6.0",
-    "eslint-plugin-import": "^2.7.0",
+    "eslint-plugin-import": "^2.11.0",
     "eslint-plugin-jsx-a11y": "^6.0.2",
     "eslint-plugin-prettier": "^2.5.0",
     "eslint-plugin-react": "^7.3.0",
@@ -34,21 +34,21 @@
     "html-loader": "^0.5.5",
     "jest": "^22.4.3",
     "loader-utils": "^1.1.0",
-    "node-sass": "^4.8.3",
+    "node-sass": "^4.9.0",
     "nodemon": "^1.14.11",
     "null-loader": "^0.1.0",
-    "postcss-loader": "^2.1.3",
-    "prettier": "^1.10.2",
-    "react-hot-loader": "^3.1.1",
+    "postcss-loader": "^2.1.4",
+    "prettier": "^1.12.1",
+    "react-hot-loader": "^4.1.2",
     "react-proxy-loader": "^0.3.5",
     "rimraf": "^2.6.1",
-    "sass-loader": "^6.0.6",
-    "source-map-support": "^0.5.2",
-    "style-loader": "^0.20.1",
+    "sass-loader": "^7.0.1",
+    "source-map-support": "^0.5.5",
+    "style-loader": "^0.21.0",
     "url-loader": "^1.0.1",
-    "webpack": "^4.2.0",
-    "webpack-cli": "^2.0.13",
-    "webpack-dev-server": "^3.1.1"
+    "webpack": "^4.6.0",
+    "webpack-cli": "^2.0.15",
+    "webpack-dev-server": "^3.1.3"
   },
   "dependencies": {
     "babel-runtime": "^6.26.0",
@@ -58,15 +58,15 @@
     "dotenv": "^5.0.0",
     "ejs": "^2.5.7",
     "express": "^4.15.4",
-    "mini-css-extract-plugin": "^0.3.0",
-    "prop-types": "^15.6.0",
+    "mini-css-extract-plugin": "^0.4.0",
+    "prop-types": "^15.6.1",
     "raf": "^3.3.2",
-    "react": "^16.0.0",
-    "react-dom": "^16.0.0",
-    "router1": "^0.11.0",
-    "router1-react": "^0.8.0",
-    "rx-react-container": "^0.5.0",
-    "rxjs": "^5.5.6"
+    "react": "^16.3.2",
+    "react-dom": "^16.3.2",
+    "router1": "^0.12.0",
+    "router1-react": "^0.8.1",
+    "rx-react-container": "^0.6.0",
+    "rxjs": "^6.0.0"
   },
   "prettier": {
     "singleQuote": true,
